Update cvs template structure, add new template, add new translations and new personal Info inputs
