+++ i18n.js
@@ -33,6 +33,8 @@ const resources = {
         level: 'Language level',
         institution: 'Institution name',
         studiesTitle: 'Your studies',
+        workPosition: 'Position name',
+        aboutMe: 'Relevant information about you',
       },
       studiesTitle: 'Your studies',
       workExperienceTitle: 'Work experience',
@@ -88,6 +90,17 @@ const resources = {
       clickMe: 'Click here!',
       infoMessage:
         'If there are any fields that you cannot fill in, leave them empty, they will not appear on your resume!',
+      workExperienceTooltip:
+        'You must add your most recent experience first, following the chronological order.',
+      selectResume: 'Select a resume template',
+      harvard: 'Harvard',
+      modern: 'Modern',
+      elegant: 'Elegant',
+      loading: 'Loading resume...',
+      selectTemplate: 'Select a resume style',
+      contact: 'Contact',
+      workPosition: 'Work position',
+      aboutMe: 'About me',
     },
   },
   es: {
@@ -121,6 +134,8 @@ const resources = {
         language: 'Nombre del idioma',
         level: 'Nivel de idioma',
         institution: 'Nombre de la institución',
+        workPosition: 'Nombre del puesto',
+        aboutMe: 'Información relevante sobre ti',
       },
       workExperienceTitle: 'Experiencia laboral',
       companyName: 'Empresa',
@@ -175,6 +190,17 @@ const resources = {
       clickMe: 'Pulsa aquí!',
       infoMessage:
         'Si hay campos que no puedes rellenar, déjalos vacíos, no aparecerán en tu CV!',
+      workExperienceTooltip:
+        'Debes añadir tu experiencia más reciente en primer lugar, siguiendo el orden cronológico.',
+      selectResume: 'Selecciona una plantilla de currículum',
+      harvard: 'Harvard',
+      modern: 'Moderno',
+      elegant: 'Elegante',
+      loading: 'Cargando CV...',
+      selectTemplate: 'Selecciona un estilo de currículum',
+      contact: 'Contacto',
+      workPosition: 'Puesto de trabajo',
+      aboutMe: 'Sobre mí',
     },
   },
 };
